reinitialized
