+++ .gitignore
@@ -9,10 +9,9 @@ npm-debug.*
 *.mobileprovision
 *.orig.*
 web-build/
-.env
+
 # macOS
 .DS_Store
 
 # Temporary files created by Metro to check the health of the file watcher
 .metro-health-check*
-eas.json
