--- app/style/color.js
@@ -1,10 +1,0 @@
-export default {
-    mainBackground: '#F8F8F8',
-    accent: '#4169e1',
-    topAccent: '#BF54F6',
-    accentLighten1: '#F9E6F5',
-    white: '#FFFFFF',
-    black: '#000000',
-    transparent: 'transparent',
-    faintBlack: '#00000080',
-}
