+++ babel.config.js
@@ -1,18 +1,6 @@
-module.exports = function (api) {
+module.exports = function(api) {
   api.cache(true);
   return {
     presets: ['babel-preset-expo'],
-    plugins: [
-      [
-        "module:react-native-dotenv", {
-          "moduleName": "@env",
-          "path": ".env",
-          "blacklist": null,
-          "whitelist": null,
-          "safe": false,
-          "allowUndefined": true
-        }
-      ]
-    ]
   };
 };
