+++ package.json
@@ -1,6 +1,6 @@
 {
-  "name": "Invoice",
-  "version": "2.0.0",
+  "name": "recido",
+  "version": "1.0.0",
   "main": "node_modules/expo/AppEntry.js",
   "scripts": {
     "start": "expo start",
@@ -9,31 +9,10 @@
     "web": "expo start --web"
   },
   "dependencies": {
-    "@expo/html-elements": "^0.4.1",
-    "@react-navigation/material-bottom-tabs": "^6.2.15",
-    "@react-navigation/material-top-tabs": "^6.6.2",
-    "@react-navigation/native": "^6.1.6",
-    "@react-navigation/stack": "^6.3.16",
-    "@reduxjs/toolkit": "1.9.3",
-    "expo": "^48.0.6",
-    "expo-print": "~12.2.1",
-    "expo-sharing": "~11.2.2",
+    "expo": "~48.0.11",
     "expo-status-bar": "~1.4.4",
-    "firebase": "^9.18.0",
-    "install": "^0.13.0",
-    "npx": "^10.2.2",
     "react": "18.2.0",
-    "react-native": "0.71.3",
-    "react-native-dotenv": "3.3.1",
-    "react-native-gesture-handler": "~2.9.0",
-    "react-native-pager-view": "6.1.2",
-    "react-native-paper": "^5.4.1",
-    "react-native-safe-area-context": "4.5.0",
-    "react-native-screens": "~3.20.0",
-    "react-native-tab-view": "^3.5.1",
-    "react-native-vector-icons": "^9.2.0",
-    "react-native-webview": "11.26.0",
-    "react-redux": "8.0.5"
+    "react-native": "0.71.6"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
